+++ tests/Jroc.Test262.Tests/built-ins/Set/prototype/values/ExecutionTests.cs
@@ -38,4 +38,7 @@ public class ExecutionTests : DiskExecutionTestsBase
     public Task length()
         => ExecutionTestFromFile("length");
 
+    [Fact(DisplayName = "name")]
+    public Task name()
+        => ExecutionTestFromFile("name");
 }